test: split real-disk I/O cases out of agent-type-manifest-loader.unit.test.ts (T-076)
The 2 cases that construct AgentTypeRegistry with no injected read/list fall
through to defaultManifestReader/defaultTypeDirLister and hit the real
filesystem — a genuine I/O boundary the .unit.test.ts suffix promises none
of. Move them to a new agent-type-manifest-loader.integration.test.ts
sibling; no logic changes, 12 injected-DI cases remain in the unit file.
